requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
PyPDF2
openai
